Further restructure feeds folder and move to app_body/views.
