+++ lib/pages/home/app_body/home_app_body.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:school_erp/features/auth/auth_repository/schemas/user.dart';
 import 'package:school_erp/pages/common_widgets/views/comming_soon_widget.dart';
 import 'package:school_erp/pages/home/app_body/views/home/home_widget.dart';
-import 'package:school_erp/pages/home/app_body/views/home/widgets/feeds.dart';
+import 'package:school_erp/pages/home/app_body/views/feeds/feeds.dart';
 import 'package:school_erp/pages/home/app_body/views/settings/settings.dart';
 
 class HomeAppBody extends StatelessWidget {
