+++ lib/pages/home/app_body/views/feeds/widgets/post_feed_modal.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/material.dart';
+
+class PostFeedModal extends StatelessWidget {
+  const PostFeedModal({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Positioned(
+      bottom: 16,
+      right: 16,
+      child: FloatingActionButton(
+        onPressed: () {
+          _showAddPostDialog(context);
+        },
+        child: Icon(Icons.add),
+      ),
+    );
+  }
+}
+
+void _showAddPostDialog(BuildContext context) {
+  showDialog(
+    context: context,
+    builder: (BuildContext context) {
+      return AlertDialog(
+        title: Text('Add New Post'),
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            TextField(
+              decoration: InputDecoration(labelText: 'Enter your post'),
+            ),
+          ],
+        ),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: Text('Cancel'),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: Text('Post'),
+          ),
+        ],
+      );
+    },
+  );
+}
